+++ Reach/test/utopia/reach/test/interactive/ReachSwitchTest.scala
@@ -4,7 +4,7 @@ import utopia.firmament.localization.LocalizedString
 import utopia.firmament.model.enumeration.StackLayout.{Center, Leading, Trailing}
 import utopia.flow.event.model.ChangeResponse.Continue
 import utopia.flow.time.TimeExtensions._
-import utopia.flow.view.mutable.eventful.EventfulPointer
+import utopia.flow.view.mutable.eventful.ResettableFlag
 import utopia.paradigm.enumeration.Alignment
 import utopia.reach.component.factory.Mixed
 import utopia.reach.component.interactive.input.check.{ContextualSwitchFactory, Switch}
@@ -30,7 +30,7 @@ object ReachSwitchTest extends App
 	import ReachTestContext._
 	
 	// Controls
-	val enabledPointer = EventfulPointer(true)
+	val enabledPointer = ResettableFlag(initialValue = true)
 	
 	// Creates the components
 	// Window